Only copy firebase plist if it doesn't exist
On CI we'll ensure the file exists before build start

+++ scripts/build-phases/generate-credentials.sh
@@ -10,7 +10,7 @@ LOCAL_SECRETS_FILE="${SRCROOT}/podcasts/Credentials/LocalApiCredentials.swift"
 LOCAL_FIREBASE_PLIST="${SRCROOT}/podcasts/Credentials/mock-GoogleService-Info.plist"
 CREDS_OUTPUT_PATH=${DERIVED_PATH}/ApiCredentials.swift
 
-FIREBASE_OUTPUT_PATH=${DERIVED_PATH}/GoogleService-Info.plist
+FIREBASE_OUTPUT_PATH=${SOURCE_ROOT}/podcasts/GoogleService-Info.plist
 
 # If the developer has a local secrets file, use it
 if [ -f "$LOCAL_SECRETS_FILE" ]; then
@@ -39,8 +39,11 @@ else
     echo ">> Generating Credentials ${CREDS_OUTPUT_PATH}"
     ruby ${SCRIPT_PATH} -i ${CREDS_INPUT_PATH} -s ${SECRETS_PATH} > "${CREDS_OUTPUT_PATH}"
 
-    ## Copy private GoogleService-Info.plist
+    ## Copy private GoogleService-Info.plist if it doesn't exist
     ##
-    echo ">> Copying Firebase Credentials from ${FIREBASE_SECRETS_PATH}"
-    echo $(<${FIREBASE_SECRETS_PATH}) > "${FIREBASE_OUTPUT_PATH}"
+    if [ ! -f $SECRETS_PATH ]; then
+        echo ">> Firebase plist not found, copying it..."
+        echo $(<${FIREBASE_SECRETS_PATH}) > "${FIREBASE_OUTPUT_PATH}"
+    fi
+    
 fi
